create unittest for each problem
