+++ csharp-project-euler-100/net.talaatharb.projecteuler100.Tests/UnitTestP001.cs
@@ -1,0 +1,13 @@
+﻿using net.talaatharb.projecteuler100.Core;
+
+namespace net.talaatharb.projecteuler100.Tests;
+
+public class UnitTestP001
+{
+    [Fact]
+    public void TestP001()
+    {
+        const long expectedResult = 233168;
+        Tester.SolveAndTest(expectedResult, new Core.P001());
+    }
+}
