+++ csharp-project-euler-100/net.talaatharb.projecteuler100.Tests/UnitTestP003.cs
@@ -1,0 +1,16 @@
+﻿using net.talaatharb.projecteuler100.Core;
+
+namespace net.talaatharb.projecteuler100.Tests;
+
+public class UnitTestP003
+{
+    /**
+    * Testing P003
+    */
+    [Fact]
+    public void TestP003()
+    {
+        const long expectedResult = 6857;
+        Tester.SolveAndTest(expectedResult, new P003());
+    }
+}
