+++ csharp-project-euler-100/net.talaatharb.projecteuler100.Tests/UnitTestP004.cs
@@ -1,0 +1,16 @@
+﻿using net.talaatharb.projecteuler100.Core;
+
+namespace net.talaatharb.projecteuler100.Tests;
+
+public class UnitTestP004
+{
+    /**
+	    * Testing P004
+	    */
+    [Fact]
+    public void TestP004()
+    {
+        const long expectedResult = 906609;
+        Tester.SolveAndTest(expectedResult, new P004());
+    }
+}
