+++ csharp-project-euler-100/net.talaatharb.projecteuler100.Tests/UnitTestP009.cs
@@ -1,0 +1,16 @@
+﻿using net.talaatharb.projecteuler100.Core;
+
+namespace net.talaatharb.projecteuler100.Tests;
+
+public class UnitTestP009
+{
+    /**
+     * Testing P009
+     */
+    [Fact]
+    public void TestP009()
+    {
+        const long expectedResult = 31875000;
+        Tester.SolveAndTest(expectedResult, new P009());
+    }
+}
